feat: Complete MVP implementation for Operational Resilience Command Center
- Implement comprehensive backend service registry and BIA models
- Build robust recovery prioritization engine integrating dependency, vendor, and risk factors
- Create incident simulator with multiple dynamic scenarios
- Develop executive brief generator with automated risk assessments
- Add extensive seed data representing an enterprise environment
- Build frontend interactive dashboard and detailed component pages
- Ensure frontend and backend are fully typed, tested, and passing CI validation
- Remove tracked build artifacts
- Add API testing and documentation

Author: Eng. Sultan Albuqami <sultan_mutep@hotmail.com>

+++ backend/app/models/bia.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, ForeignKey, Integer, String
+from sqlalchemy import Column, Float, ForeignKey, Integer, String
 from sqlalchemy.orm import relationship
 
 from app.db.base_class import Base
@@ -10,7 +10,7 @@ class BIA(Base):
     id = Column(Integer, primary_key=True, index=True)
     service_id = Column(Integer, ForeignKey("services.id"), nullable=False, unique=True)
 
-    rto_target_hours = Column(Integer, nullable=False)
+    rto_target_hours = Column(Float, nullable=False)
     rpo_target_minutes = Column(Integer, nullable=False)
 
     financial_impact = Column(String, nullable=False)
